requirements.txt
streamlit
supabase
pandas
python-dotenv
plotly
altair
google-generativeai
Pillow
